unsigned attrlen

+++ p4/insert.cpp
@@ -37,7 +37,7 @@ Status Updates::Insert(const string& relation,      // Name of the relation
     for (int i = 0; i < attrCnt; i++)
     {
         AttrDesc currAttr = attrDesc[attrMap[const_cast<char*>(attrList[i].attrName)]];
-        if ((Datatype) attrList[i].attrType == STRING && strlen((char*) attrList[i].attrValue) != attrList[i].attrLen)
+        if ((Datatype) attrList[i].attrType == STRING && strlen((char*) attrList[i].attrValue) != (unsigned int) attrList[i].attrLen)
         {
             return ATTRTOOLONG;
         }
